refactor, fuzz: Split miniscript utils from harness
